[B]ADD_SearchDetail

+++ backend/src/main/java/com/jagi/yeobo/domain/repository/BagRepository.java
@@ -21,17 +21,17 @@ public class BagRepository {
     @PersistenceContext
     private final EntityManager em;
 
-    public Bag findByBag(int bagId){
+    public Bag findByBag(long bagId){
         Bag findBag = em.find(Bag.class, bagId);
         return findBag;
     }
 
-    public void updateBag(int bagId, BagDto bagDto){
+    public void updateBag(long bagId, BagDto bagDto){
         Bag findBag = findByBag(bagId);
         findBag.updateBag(bagDto);
     }
 
-    public List<BagDto> searchBagList(int userId){
+    public List<BagDto> searchBagList(long userId){
         List<Bag> bagList = em.createQuery("SELECT b FROM bag as b WHERE b.user_id = :userId", Bag.class)
                 .setParameter("userId", userId).getResultList();
         List<BagDto> bagDtoList = new ArrayList<>();
@@ -45,7 +45,7 @@ public class BagRepository {
         return bagDtoList;
      }
 
-    public void likeBag(int userId, int bagId){
+    public void likeBag(long userId, long bagId){
         Bag findBag = em.find(Bag.class, bagId);
 
         int currentCnt = findBag.getLikeCnt();
@@ -59,7 +59,7 @@ public class BagRepository {
         em.persist(newPick);
     }
 
-    public List<BagDto> searchLikeBagList(int userId){
+    public List<BagDto> searchLikeBagList(long userId){
         List<Pick> pickList = em.createQuery("SELECT p From pick as p WHERE p.user_id = :userId", Pick.class)
                 .setParameter("userId", userId).getResultList();
